rename plugins to services
