+++ server/config.go
@@ -24,10 +24,10 @@ import (
 	"go.woodpecker-ci.org/woodpecker/v2/server/forge"
 	"go.woodpecker-ci.org/woodpecker/v2/server/logging"
 	"go.woodpecker-ci.org/woodpecker/v2/server/model"
-	"go.woodpecker-ci.org/woodpecker/v2/server/plugins"
-	"go.woodpecker-ci.org/woodpecker/v2/server/plugins/permissions"
 	"go.woodpecker-ci.org/woodpecker/v2/server/pubsub"
 	"go.woodpecker-ci.org/woodpecker/v2/server/queue"
+	"go.woodpecker-ci.org/woodpecker/v2/server/services"
+	"go.woodpecker-ci.org/woodpecker/v2/server/services/permissions"
 )
 
 var Config = struct {
@@ -37,7 +37,7 @@ var Config = struct {
 		Logs       logging.Log
 		Forge      forge.Forge
 		Membership cache.MembershipService
-		Manager    *plugins.Manager
+		Manager    *services.Manager
 	}
 	Storage struct {
 		// Users  model.UserStore
