+++ server/services/config/forge_test.go
@@ -16,25 +16,18 @@ package config_test
 
 import (
 	"context"
-	"crypto/ed25519"
-	"crypto/rand"
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
-	"net/http/httptest"
 	"path/filepath"
 	"testing"
 	"time"
 
-	"github.com/go-ap/httpsig"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
 	"go.woodpecker-ci.org/woodpecker/v2/server/forge/mocks"
 	forge_types "go.woodpecker-ci.org/woodpecker/v2/server/forge/types"
 	"go.woodpecker-ci.org/woodpecker/v2/server/model"
-	"go.woodpecker-ci.org/woodpecker/v2/server/plugins/config"
+	"go.woodpecker-ci.org/woodpecker/v2/server/services/config"
 )
 
 func TestFetch(t *testing.T) {
@@ -315,213 +308,6 @@ func TestFetch(t *testing.T) {
 			configFetcher := config.NewForge(
 				time.Second * 3,
 			)
-			files, err := configFetcher.Fetch(
-				context.Background(),
-				f,
-				&model.User{Token: "xxx"},
-				repo,
-				&model.Pipeline{Commit: "89ab7b2d6bfb347144ac7c557e638ab402848fee"},
-			)
-			if tt.expectedError && err == nil {
-				t.Fatal("expected an error")
-			} else if !tt.expectedError && err != nil {
-				t.Fatal("error fetching config:", err)
-			}
-
-			matchingFiles := make([]string, len(files))
-			for i := range files {
-				matchingFiles[i] = files[i].Name
-			}
-			assert.ElementsMatch(t, tt.expectedFileNames, matchingFiles, "expected some other pipeline files")
-		})
-	}
-}
-
-func TestFetchFromConfigService(t *testing.T) {
-	t.Parallel()
-
-	type file struct {
-		name string
-		data []byte
-	}
-
-	dummyData := []byte("TEST")
-
-	testTable := []struct {
-		name              string
-		repoConfig        string
-		files             []file
-		expectedFileNames []string
-		expectedError     bool
-	}{
-		{
-			name:              "External Fetch empty repo",
-			repoConfig:        "",
-			files:             []file{},
-			expectedFileNames: []string{"override1", "override2", "override3"},
-			expectedError:     false,
-		},
-		{
-			name:       "Default config - Additional sub-folders",
-			repoConfig: "",
-			files: []file{{
-				name: ".woodpecker/test.yml",
-				data: dummyData,
-			}, {
-				name: ".woodpecker/sub-folder/config.yml",
-				data: dummyData,
-			}},
-			expectedFileNames: []string{"override1", "override2", "override3"},
-			expectedError:     false,
-		},
-		{
-			name:       "Fetch empty",
-			repoConfig: " ",
-			files: []file{{
-				name: ".woodpecker/.keep",
-				data: dummyData,
-			}, {
-				name: ".woodpecker.yml",
-				data: nil,
-			}, {
-				name: ".woodpecker.yaml",
-				data: dummyData,
-			}},
-			expectedFileNames: []string{},
-			expectedError:     true,
-		},
-		{
-			name:       "Use old config",
-			repoConfig: ".my-ci-folder/",
-			files: []file{{
-				name: ".woodpecker/test.yml",
-				data: dummyData,
-			}, {
-				name: ".woodpecker.yml",
-				data: dummyData,
-			}, {
-				name: ".woodpecker.yaml",
-				data: dummyData,
-			}, {
-				name: ".my-ci-folder/test.yml",
-				data: dummyData,
-			}},
-			expectedFileNames: []string{
-				".my-ci-folder/test.yml",
-			},
-			expectedError: false,
-		},
-	}
-
-	pubEd25519Key, privEd25519Key, err := ed25519.GenerateKey(rand.Reader)
-	if err != nil {
-		t.Fatal("can't generate ed25519 key pair")
-	}
-
-	fixtureHandler := func(w http.ResponseWriter, r *http.Request) {
-		// check signature
-		pubKeyID := "woodpecker-ci-plugins"
-
-		keystore := httpsig.NewMemoryKeyStore()
-		keystore.SetKey(pubKeyID, pubEd25519Key)
-
-		verifier := httpsig.NewVerifier(keystore)
-		verifier.SetRequiredHeaders([]string{"(request-target)", "date"})
-
-		keyID, err := verifier.Verify(r)
-		if err != nil {
-			http.Error(w, "Invalid signature", http.StatusBadRequest)
-			return
-		}
-
-		if keyID != pubKeyID {
-			http.Error(w, "Used wrong key", http.StatusBadRequest)
-			return
-		}
-
-		type config struct {
-			Name string `json:"name"`
-			Data string `json:"data"`
-		}
-
-		type incoming struct {
-			Repo          *model.Repo     `json:"repo"`
-			Build         *model.Pipeline `json:"pipeline"`
-			Configuration []*config       `json:"config"`
-		}
-
-		var req incoming
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, "can't read body", http.StatusBadRequest)
-			return
-		}
-		err = json.Unmarshal(body, &req)
-		if err != nil {
-			http.Error(w, "Failed to parse JSON"+err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		if req.Repo.Name == "Fetch empty" {
-			w.WriteHeader(404)
-			return
-		}
-
-		if req.Repo.Name == "Use old config" {
-			w.WriteHeader(204)
-			return
-		}
-
-		fmt.Fprint(w, `{
-			"configs": [
-					{
-							"name": "override1",
-							"data": "some new pipelineconfig \n pipe, pipe, pipe"
-					},
-					{
-							"name": "override2",
-							"data": "some new pipelineconfig \n pipe, pipe, pipe"
-					},
-					{
-							"name": "override3",
-							"data": "some new pipelineconfig \n pipe, pipe, pipe"
-					}
-			]
-}`)
-	}
-
-	ts := httptest.NewServer(http.HandlerFunc(fixtureHandler))
-	defer ts.Close()
-	configAPI := config.NewHTTP(ts.URL, privEd25519Key)
-
-	for _, tt := range testTable {
-		t.Run(tt.name, func(t *testing.T) {
-			repo := &model.Repo{Owner: "laszlocph", Name: tt.name, Config: tt.repoConfig} // Using test name as repo name to provide different responses in mock server
-
-			f := new(mocks.Forge)
-			dirs := map[string][]*forge_types.FileMeta{}
-			for _, file := range tt.files {
-				f.On("File", mock.Anything, mock.Anything, mock.Anything, mock.Anything, file.name).Return(file.data, nil)
-				path := filepath.Dir(file.name)
-				if path != "." {
-					dirs[path] = append(dirs[path], &forge_types.FileMeta{
-						Name: file.name,
-						Data: file.data,
-					})
-				}
-			}
-
-			for path, files := range dirs {
-				f.On("Dir", mock.Anything, mock.Anything, mock.Anything, mock.Anything, path).Return(files, nil)
-			}
-
-			// if the previous mocks do not match return not found errors
-			f.On("File", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil, fmt.Errorf("file not found"))
-			f.On("Dir", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil, fmt.Errorf("directory not found"))
-
-			f.On("Netrc", mock.Anything, mock.Anything).Return(&model.Netrc{Machine: "mock", Login: "mock", Password: "mock"}, nil)
-
-			configFetcher := config.NewForge(time.Second * 3)
 			files, err := configFetcher.Fetch(
 				context.Background(),
 				f,
