+++ server/services/registry/combined.go
