+++ server/services/registry/service.go
